# src/product-service/docs/swagger.yaml
definitions:
  common.AppError:
    properties:
      key:
        type: string
      log:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      filter: {}
      message: {}
      paging: {}
    type: object
...
  models.ProductWarehouses:
    properties:
      last_updated:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
      warehouse_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/product/warehouse:
    post:
      consumes:
      - application/json
      description: Create a new product warehouse with the provided parameters
      parameters:
      - description: Product warehouse to create
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.ProductWarehouses'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Create a new product warehouse
      tags:
      - product warehouse
    put:
      consumes:
      - application/json
      description: Update an existing product warehouse with the provided parameters
      parameters:
      - description: Product warehouse to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.ProductWarehouses'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Update an existing product warehouse
      tags:
      - product warehouse
  /api/product/warehouse/{idProduct}/{idWarehouse}:
    delete:
      consumes:
      - application/json
      description: Delete a product warehouse by id
      parameters:
      - description: Product ID
        in: path
        name: idProduct
        required: true
        type: string
      - description: Warehouse ID
        in: path
        name: idWarehouse
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Deleted a product warehouse
      tags:
      - product warehouse
    get:
      consumes:
      - application/json
      description: Get a product warehouse
      parameters:
      - description: Product ID
        in: path
        name: idProduct
        required: true
        type: string
      - description: Warehouse ID
        in: path
        name: idWarehouse
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Get a product warehouse
      tags:
      - product warehouse
  /api/product/warehouses:
    get:
      consumes:
      - application/json
      description: Retrieve all product warehouses, with optional paging
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: '-'
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.AppError'
      summary: Retrieve a list of product warehouses
      tags:
      - product warehouse
...
swagger: "2.0"